pyproject.toml
[tool.poetry]
name = "test-task"
version = "0.1.0"
description = ""
authors = ["alekseybeloded <belod-aleksej@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.1"
djangorestframework = "^3.15.2"
channels = {extras = ["daphne"], version = "^4.1.0"}
channels-redis = "^4.2.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
ruff = "^0.6.9"

[tool.ruff]
# Enable the pycodestyle (`E`) and Pyflakes (`F`) rules by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
lint.select = ["E", "E3", "F", "B", "I", "W", "UP", "T20"]
lint.ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
lint.fixable = ["ALL"]
lint.unfixable = []
lint.preview = true

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "migrations",
    "settings.py",
    "test_*.py",
    "conftest.py",
    ]

line-length = 100

# Allow unused variables when underscore-prefixed.
lint.dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Assume Python 3.11
target-version = 'py311'


[tool.ruff.format]
# 5. Use single quotes for non-triple-quoted strings.
quote-style = "single"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
